Reorganize package: amici.exporters.sundials
Move modules and templates related to generating C++ code for simulation with sundials solvers to `amici.exporters.sundials`.
