+++ python/sdist/amici/exporters/sundials/templates/CMakeLists.template.cmake
