+++ python/sdist/amici/exporters/sundials/templates/MANIFEST.template.in
