+++ python/sdist/amici/exporters/sundials/templates/setup.template.py
@@ -5,11 +5,10 @@ import os
 import sys
 from pathlib import Path
 
-from cmake_build_extension import CMakeExtension
-from setuptools import find_namespace_packages, setup
-
 from amici import _get_amici_path
 from amici.custom_commands import AmiciBuildCMakeExtension
+from cmake_build_extension import CMakeExtension
+from setuptools import find_namespace_packages, setup
 
 
 def get_extension() -> CMakeExtension:
