+++ python/tests/test_ode_export.py
@@ -1,7 +1,7 @@
 """Miscellaneous AMICI Python interface tests"""
 
 import sympy as sp
-from amici.cxxcodeprinter import csc_matrix
+from amici.exporters.sundials.cxxcodeprinter import csc_matrix
 from amici.import_utils import symbol_with_assumptions
 from amici.testing import skip_on_valgrind
